Joto class working

+++ tests/test_joto.py
@@ -21,15 +21,17 @@ import joto
 class TestSQLiteDB(unittest.TestCase):
 
     def test_create_table(self):
+        print("----------------------")
+        print("Test creating db")
+        print("----------------------")
 
         db_path = "test.db"
-        table = '''test'''
 
         if os.path.isfile(db_path): 
             os.remove(db_path)
             print("First remove exiting db")
 
-        sqlite_db = joto.SQLiteDB(db_path,table)
+        sqlite_db = joto.JotoSQLiteDB(db_path)
         sqlite_db.create_db_table()
 
         if os.path.isfile(db_path):
@@ -37,16 +39,18 @@ class TestSQLiteDB(unittest.TestCase):
 
         self.assertTrue(result) 
 
-    def test_add_retrieve_data(self):
+    def test_add_and_get_data(self):
+        print("----------------------")
+        print("Test getting data from table db")
+        print("----------------------")
 
         db_path = "test.db"
-        table = '''test'''
 
         if os.path.isfile(db_path): 
             os.remove(db_path)
-            print("First remove exiting db")
+            print("First remove existing db")
 
-        sqlite_db = joto.SQLiteDB(db_path,table)
+        sqlite_db = joto.JotoSQLiteDB(db_path)
         sqlite_db.create_db_table()
         sqlite_db.check_for_table()
 
@@ -63,31 +67,36 @@ class TestSQLiteDB(unittest.TestCase):
         self.assertTrue(result) 
 
     def test_compress_image(self):
-
-        size_reduction_factor = 10 #%
+        print("----------------------")
+        print("Test image compression")
+        print("----------------------")
 
         image = "2020-03-17_Sox.jpg"
-        src_dir = "ingest_images/"
-        dst_dir = "compressed_images/"
-        achv_dir = "archived_images/"
         size = "1000x1000"
+        size_reduction_factor = 10 #%
+
+        #setup
+        #--------------------------------------------------
+        db_path = "test.db"
+        src_dir = "ingest_images/"
+        dst_dir = "images/compressed/"
+        achv_dir = "images/original/"
+
+        if os.path.exists(db_path): os.remove(db_path)
+        if os.path.exists(src_dir): shutil.rmtree(src_dir)
+        if os.path.exists(dst_dir): shutil.rmtree(dst_dir)
+        if os.path.exists(achv_dir): shutil.rmtree(achv_dir)
+
+        os.mkdir(src_dir) 
+        os.makedirs(dst_dir) 
+        os.makedirs(achv_dir) 
+        shutil.copy(image,src_dir + image)
+        #--------------------------------------------------
 
         original_image_size = int(os.stat(image)[6])
 
-        #setup
-        try:
-            shutil.rmtree(src_dir)
-            shutil.rmtree(dst_dir)
-            shutil.rmtree(achv_dir)
-        except: pass
-
-        os.mkdir(src_dir) 
-        os.mkdir(dst_dir) 
-        os.mkdir(achv_dir) 
-        shutil.copy(image,src_dir + image)
-
         images_manage = joto.ImagesManage(size,src_dir,dst_dir,achv_dir)
-        images_manage.compress_image(image)
+        images_manage.compress_and_archive_image(image)
 
         compressed_image_size = int(os.stat(dst_dir + image)[6])
 
@@ -95,9 +104,76 @@ class TestSQLiteDB(unittest.TestCase):
             result = True
 
         self.assertTrue(result)
-        
 
+    def test_create_requirements(self):
+        print("------------------------------------------")
+        print("Integration test: creating paths, db, table")
+        print("------------------------------------------")
 
+        image = "2020-03-17_Sox.jpg"
+        size = "1000x1000"
+
+        #setup
+        #--------------------------------------------------
+        db_path = "test.db"
+        src_dir = "ingest_images/"
+        dst_dir = "images/compressed/"
+        achv_dir = "images/original/"
+
+        if os.path.exists(db_path): os.remove(db_path)
+        if os.path.exists(src_dir): shutil.rmtree(src_dir)
+        if os.path.exists(dst_dir): shutil.rmtree(dst_dir)
+        if os.path.exists(achv_dir): shutil.rmtree(achv_dir)
+
+        # shutil.copy(image,src_dir + image)
+        #--------------------------------------------------
+
+        sqlite_db = joto.JotoSQLiteDB(db_path)
+        images_manage = joto.ImagesManage(size, src_dir, dst_dir, achv_dir)
+        joto_obj = joto.Joto(sqlite_db,images_manage)
+        joto_obj.create_all_requirements()
+        joto_obj.check_requirements()
+        # joto_obj.scan_for_and_add_images_with_text()
+
+        self.assertTrue(True)
+
+    def test_adding_data(self):
+        print("------------------------------------------")
+        print("Integration test: adding image")
+        print("------------------------------------------")
+
+        image = "2020-03-17_Sox.jpg"
+        size = "1000x1000"
+
+        #setup
+        #--------------------------------------------------
+        db_path = "test.db"
+        src_dir = "ingest_images/"
+        dst_dir = "images/compressed/"
+        achv_dir = "images/original/"
+
+        if os.path.exists(db_path): os.remove(db_path)
+        if os.path.exists(src_dir): shutil.rmtree(src_dir)
+        if os.path.exists(dst_dir): shutil.rmtree(dst_dir)
+        if os.path.exists(achv_dir): shutil.rmtree(achv_dir)
+
+        #--------------------------------------------------
+
+        sqlite_db = joto.JotoSQLiteDB(db_path)
+        images_manage = joto.ImagesManage(size, src_dir, dst_dir, achv_dir)
+
+        joto_obj = joto.Joto(sqlite_db,images_manage)
+        joto_obj.create_all_requirements()
+        joto_obj.check_requirements()
+
+        shutil.copy(image,src_dir + image)
+        joto_obj.scan_for_and_add_images_with_text()
+        all_data = sqlite_db.retrieve_all_data_ordered_by_date()
+        if all_data[0][4] == image:
+            result = True
+        # joto_obj.scan_for_and_add_images_with_text()
+
+        self.assertTrue(result)
 
 if __name__ == '__main__':
     unittest.main()
